LayoutTests/fast/dom/Document/adoptNode-null.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests that document.adoptNode(null) throws a TypeError");

shouldThrow("document.adoptNode(null)", "'TypeError: Argument 1 (\\'source\\') to Document.adoptNode must be an instance of Node'");
shouldThrow("document.adoptNode()", "'TypeError: Not enough arguments'");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
